flask_dynamic_secrets_app/requirements.txt
Flask==3.0.2
requests
Werkzeug>=3.0.0
mysql-connector-python==8.3.0
watchdog
gunicorn==21.0.1
